出し物にズームする機能の追加
